+++ website/module.f.ts
@@ -1,11 +1,12 @@
-import type { Io } from '../io/module.f.ts'
-import { encodeUtf8 } from '../types/uint8array/module.f.ts'
 import { htmlToString, type Element } from '../html/module.f.ts'
+import { type NodeOperations, writeFile, type NodeProgram } from '../types/effect/node/module.f.ts'
+import { utf8 } from '../text/module.f.ts'
 
-export const run = async({ fs: { promises: { writeFile }} }: Io): Promise<number> => {
-    const html: Element = ['body',
-        ['a', { href: 'https://github.com/functionalscript/functionalscript' }, 'GitHub Repository']
-    ]
-    await writeFile('index.html', encodeUtf8(htmlToString(html)))
-    return 0
-}
+const html: Element = ['body',
+    ['a', { href: 'https://github.com/functionalscript/functionalscript' }, 'GitHub Repository']
+]
+
+const program = writeFile<NodeOperations>('index.html', utf8(htmlToString(html)))
+    .map(() => 0)
+
+export const run: NodeProgram = () => program